d3d12, d3d12core: Refactor to move main code to D3D12Core
Move the main bulk of the code to d3d12core, and expose an interface to access.

This allows us to pass the new test that prove that the vtable is resident in d3d12core going forward.

Essentially turns d3d12 into a slim wrapper.
