+++ include/vkd3d.h
@@ -22,6 +22,7 @@
 #include <vkd3d_types.h>
 
 #ifndef VKD3D_NO_WIN32_TYPES
+# define COBJMACROS
 # include <vkd3d_windows.h>
 
 # define WIDL_C_INLINE_WRAPPERS
@@ -35,6 +36,7 @@
 # endif
 
 # include <vkd3d_d3d12.h>
+# include <vkd3d_core_interface.h>
 # undef WIDL_C_INLINE_WRAPPERS
 #endif  /* VKD3D_NO_WIN32_TYPES */
