+++ include/vkd3d_sonames.h
@@ -23,10 +23,13 @@
 
 #if defined(_WIN32)
 #define SONAME_LIBVULKAN "vulkan-1.dll"
+#define SONAME_D3D12CORE "d3d12core.dll"
 #elif defined(__linux__)
 #define SONAME_LIBVULKAN "libvulkan.so.1"
+#define SONAME_D3D12CORE "libvkd3d-proton-d3d12core.so"
 #elif defined(__APPLE__)
 #define SONAME_LIBVULKAN "libvulkan.1.dylib"
+#define SONAME_D3D12CORE "vkd3d-proton-d3d12core.dylib"
 #else
 #error "Unrecognized platform."
 #endif
