+++ libs/meson.build
@@ -2,6 +2,4 @@ subdir('vkd3d-common')
 subdir('vkd3d-shader')
 subdir('vkd3d')
 subdir('d3d12')
-if vkd3d_platform == 'windows'
-  subdir('d3d12core')
-endif
+subdir('d3d12core')
